Fixed tiles, surfaces and string functions
-Tiles broke when I "optimized" d3d_model_draw function. Fixed it in
GL3screen.
-Surfaces set_target and reset_target no correctly flushes the batch.
-Fixed segfault when calling string_lower and string_upper with empty
string. It was actually broken even if we called it correctly as the
last character wasn't converted.

+++ ENIGMAsystem/SHELL/Universal_System/estring.h
@@ -49,7 +49,7 @@ string string_delete(string str,int index,int count);
 string string_insert(string substr,string str,int index);
 string string_replace(string str,string substr,string newstr);
 string string_replace_all(string str,string substr,string newstr);
-int string_count(string substr, string str);
+size_t string_count(string substr, string str);
 
 string string_lower(string str);
 string string_upper(string str);
